Fixed super cell flag
Usage:

If one use -usesupercell, then a 2 X 2 X 2  supercell is created
that generates a new grid on which all the modules are computed

Changes:

The earlier method was based on using a vtkTransform to transform
the vtkImageData grid, but this generated an structured Grid
as output of the vtkAppendFilter. The code than had to branched if
supercell is used as there was no way to convert vtkStructureGrid
to imageData. Moreover, file sizes close
10 GB were created and huge amount of RAM was used rendering it
quite inefficient.

Now, the grid that is read as vtkImageData is simply appended
multiple times using the vtkAppendImage filter and writing
using the vtkXMLImageDataWriter. This takes very short time,
does not branch code and writes files in MB's