ADD: stress grids can also be now run as inputs
Often, once we segment nanoporous gold structures, we would also
like to visualize the stresses in the segments. For this, first
using the python script generate-stress-grids.py, first the grid
that has the distance grid and stresses is generated, which is
read as input to the tda-segmentor code. This allows
visualization of stresses in the segments